Replace all of callable operations to PyCallable
