+++ vm/src/protocol/callable.rs
@@ -1,7 +1,7 @@
 use crate::{
     function::IntoFuncArgs,
     types::GenericMethod,
-    {PyObject, PyResult, VirtualMachine}
+    {AsObject, PyObject, PyResult, VirtualMachine},
 };
 
 impl PyObject {
@@ -28,6 +28,70 @@ impl<'a> PyCallable<'a> {
     }
 
     pub fn invoke(&self, args: impl IntoFuncArgs, vm: &VirtualMachine) -> PyResult {
-        (self.call)(self.obj, args.into_args(vm), vm)
+        vm.trace_event(TraceEvent::Call)?;
+        let result = (self.call)(self.obj, args.into_args(vm), vm);
+        vm.trace_event(TraceEvent::Return)?;
+        result
+    }
+}
+
+/// Trace events for sys.settrace and sys.setprofile.
+enum TraceEvent {
+    Call,
+    Return,
+}
+
+impl std::fmt::Display for TraceEvent {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        use TraceEvent::*;
+        match self {
+            Call => write!(f, "call"),
+            Return => write!(f, "return"),
+        }
+    }
+}
+
+impl VirtualMachine {
+    /// Call registered trace function.
+    #[inline]
+    fn trace_event(&self, event: TraceEvent) -> PyResult<()> {
+        if self.use_tracing.get() {
+            self._trace_event_inner(event)
+        } else {
+            Ok(())
+        }
+    }
+    fn _trace_event_inner(&self, event: TraceEvent) -> PyResult<()> {
+        let trace_func = self.trace_func.borrow().to_owned();
+        let profile_func = self.profile_func.borrow().to_owned();
+        if self.is_none(&trace_func) && self.is_none(&profile_func) {
+            return Ok(());
+        }
+
+        let frame_ref = self.current_frame();
+        if frame_ref.is_none() {
+            return Ok(());
+        }
+
+        let frame = frame_ref.unwrap().as_object().to_owned();
+        let event = self.ctx.new_str(event.to_string()).into();
+        let args = vec![frame, event, self.ctx.none()];
+
+        // temporarily disable tracing, during the call to the
+        // tracing function itself.
+        if !self.is_none(&trace_func) {
+            self.use_tracing.set(false);
+            let res = self.invoke(&trace_func, args.clone());
+            self.use_tracing.set(true);
+            res?;
+        }
+
+        if !self.is_none(&profile_func) {
+            self.use_tracing.set(false);
+            let res = self.invoke(&profile_func, args);
+            self.use_tracing.set(true);
+            res?;
+        }
+        Ok(())
     }
 }
